refactor: 주문 완료된 메뉴 값객체로 분리
