+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -5,7 +5,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.generic.Price;
 
 @Entity
 public class OrderLineItem {
@@ -13,21 +12,19 @@ public class OrderLineItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long seq;
-    private String menuName;
+
     @Embedded
-    private Price menuPrice;
+    private OrderedMenu orderedMenu;
     private long quantity;
 
     protected OrderLineItem() {
     }
 
     public OrderLineItem(
-            final String menuName,
-            final Price menuPrice,
+            final OrderedMenu orderedMenu,
             final long quantity
     ) {
-        this.menuName = menuName;
-        this.menuPrice = menuPrice;
+        this.orderedMenu = orderedMenu;
         this.quantity = quantity;
     }
 
@@ -35,12 +32,8 @@ public class OrderLineItem {
         return seq;
     }
 
-    public String getMenuName() {
-        return menuName;
-    }
-
-    public Price getMenuPrice() {
-        return menuPrice;
+    public OrderedMenu getOrderedMenu() {
+        return orderedMenu;
     }
 
     public long getQuantity() {
